Make web server python 3 compatible
Rename `http.py` to `http-server.py` since `http.py` caused import issues in
python 3, where there is a built in `http` module.
